core: add register file
